<refactor, fix> Running Config & Editions Fix
Code from previous commit adding support to Running Configs has been
refactorized and improved.

A bug was found. Attribute "edition" was considered in the creation
of the FM even though it is officially deprecated. Hence, no CPEs
were found if a constraint was made using "edition" as filter, leading
to faulty RestrictionNodes.

Bug was fixed by deleting this attribute from the list of attributes
considered when constructing the FM.